+++ app/src/main/java/spirit/realm/faefinance/data/classes/PeriodicTransaction.kt
@@ -9,14 +9,14 @@ import java.util.Date
     ]
 )
 data class PeriodicTransaction(
-    @PrimaryKey(autoGenerate = true) val id: Int,
+    @PrimaryKey(autoGenerate = true) val id: Long,
     var type: ETransactionType,
     var title: String,
     var amount: Double,
-    var senderAccount: Int, // FK to Account.id
-    var recipientAccount: Int, // FK to Account.id
+    var senderAccount: Long, // FK to Account.id
+    var recipientAccount: Long, // FK to Account.id
     var currency: String, // ISO code
-    var category: Int, // FK to Category.id
+    var category: Long, // FK to Category.id
     var nextTransaction: Date,
     var interval: ETransactionInterval,
     var intervalLength: Int
