+++ app/src/main/java/spirit/realm/faefinance/data/classes/Transaction.kt
@@ -12,14 +12,14 @@ import java.util.Date
     ]
 )
 data class Transaction(
-    @PrimaryKey(autoGenerate = true) val id: Int,
+    @PrimaryKey(autoGenerate = true) val id: Long,
     var type: ETransactionType,
     var title: String,
     var amount: Double,
     var amountConverted: Double,
-    var senderAccount: Int, // FK to Account.id
-    var recipientAccount: Int, // FK to Account.id
+    var senderAccount: Long, // FK to Account.id
+    var recipientAccount: Long, // FK to Account.id
     var currency: String, // ISO code
-    var category: Int, // FK to Category.id
+    var category: Long, // FK to Category.id
     var timestamp: Date
 )
